+++ archive.php
@@ -28,8 +28,6 @@
 
 				<?php endif; ?>
 
-				<?php get_sidebar(); ?>
-
 			</main>
 
 <?php get_footer(); ?>