+++ author.php
@@ -41,8 +41,6 @@ get_header(); ?>
 
 				<?php endif; ?>
 
-				<?php // get_sidebar(); ?>
-
 			</main>
 
 <?php get_footer(); ?>